Title Menu
This is the Title Menu class, an extension of the GameState class. It is used in the core, our GameWindow  class, as the default game state when the application is launched. So far, it only has one box, that changes color if up or down is pressed, additional functionality should be added (such as named selections, like "start game" or "quit").