Player
This is the player class, that extends GameObject. It should be usable in a game scene, but when I tried that I ran into some performance issues. Currently, there are already some performance issues (loss of framerate) if the game runs for a little while. I'm not sure if that is in this class or not, but it's when the Player is used that it's most noticeable. One idea for fixing this is to control the rate at which frames are displayed / updates are taken.