Pause Menu
Pause Menu works, however, I have been unable to access it as I intend. I want to be able to press "p" during the game GameState, and have this Game State switched to, but it doesn't work. I'm not sure why. But if the game starts at Pause Menu, Pause Menu functions as intended.